utilities/earthly-cache-watcher/requirements.txt
watchdog
python-dotenv